+++ src/main/scala/com/atomist/rug/runtime/plans/PlanResultInterpreter.scala
@@ -21,7 +21,7 @@ object PlanResultInterpreter {
   }
 
   @tailrec
-  private def hasLogFailure(log: Seq[PlanLogEvent]): Boolean = {
+  def hasLogFailure(log: Seq[PlanLogEvent]): Boolean = {
     log.headOption match {
       case Some(head) =>
         head match {
@@ -35,5 +35,4 @@ object PlanResultInterpreter {
       case None => false
     }
   }
-
 }